Write speed command 0 if hazard is detected
This output can be connected to our path/waypoint following
component in order to stop the rover without having to
connect to the EMERGENCY_HAZARD state.

Since this is only written once (see previous commit),
the rover can now stop in front of a hazard until the path
planner finishes computing a new path, which avoids the obstacle,
and continue driving.